pyproject.toml
[tool.poetry]
name = "kgenealogic"
version = "0.1.0"
description = "Genetic genealogy tool for tree structure inference"
authors = ["John Wilmes <jw@johnwilmes.name>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.8"
scipy = "^1.10.1"
numpy = "^1.24.3"
pandas = "^2.0.1"
osqp = "^0.6.2.post9"
pysqlite3-binary = "^0.5.0"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
